+++ annotations/nocedit.exe/pseudocode/src/core/msnedit.cpp/drawMotionBar_FUN_00536e20.cpp
@@ -39,7 +39,7 @@ void __cdecl core_msnedit_cpp_drawMotionBar_FUN_00536e20(void)
     engine_2d_c_fillRectWithBorder_FUN_00403200(1,iVar2,0x9b,iVar1,0,0xff);
     motion_name = core_motion_cpp_CMotionController_getCurrentMotion_FUN_0052dab0
                             (&this_ptr->motion_controller);
-    local_f8 = (float)(this_ptr->motion_controller).current_frame_number;
+    local_f8 = (this_ptr->motion_controller).current_frame_number;
     core_motion_cpp_CMotionController_getCurrentStateName_FUN_0052dcf0(&this_ptr->motion_controller)
     ;
     _sprintf(local_f4,"%s (%s)",motion_name);
